added MakeTransaction almost done

+++ goofy_project/goofy_app/models.py
@@ -45,7 +45,7 @@ class Transaction(models.Model):
     recipient = models.ForeignKey(
         User, related_name="received_transactions", on_delete=models.CASCADE
     )
-    amount = models.DecimalField(max_digits=10, decimal_places=2)
+    amount = models.IntegerField()
     timestamp = models.DateTimeField(auto_now_add=True)
     signature = models.TextField()  # Store signature of the transaction
